+++ R/extract.R
@@ -23,11 +23,15 @@
 #' - `extract_recipe()` returns the recipe. The `estimated` argument specifies
 #'    whether the fitted or original recipe is returned.
 #'
+#' - `extract_parameter_dials()` returns a single dials parameter object.
+#'
+#' - `extract_parameter_set_dials()` returns a set of dials parameter objects.
+#'
 #' @param x A workflow
 #'
 #' @param estimated A logical for whether the original (unfit) recipe or the
 #' fitted recipe should be returned. This argument should be named.
-#'
+#' @param parameter A single string for the parameter ID.
 #' @param ... Not currently used.
 #'
 #' @details
@@ -117,7 +121,7 @@ extract_spec_parsnip.workflow <- function(x, ...) {
 #' @export
 #' @rdname extract-workflow
 extract_recipe.workflow <- function(x, ..., estimated = TRUE) {
-  ellipsis::check_dots_empty()
+  check_dots_empty()
   if (!is_bool(estimated)) {
     abort("`estimated` must be a single `TRUE` or `FALSE`.")
   }
@@ -141,7 +145,10 @@ extract_fit_parsnip.workflow <- function(x, ...) {
   if (has_fit(x)) {
     return(x$fit$fit)
   }
-  abort("The workflow does not have a model fit. Have you called `fit()` yet?")
+  abort(c(
+    "Can't extract a model fit from an untrained workflow.",
+    i = "Do you need to call `fit()`?"
+  ))
 }
 
 #' @export
@@ -156,7 +163,10 @@ extract_mold.workflow <- function(x, ...) {
   if (has_mold(x)) {
     return(x$pre$mold)
   }
-  abort("The workflow does not have a mold. Have you called `fit()` yet?")
+  abort(c(
+    "Can't extract a mold from an untrained workflow.",
+    i = "Do you need to call `fit()`?"
+  ))
 }
 
 #' @export
@@ -172,4 +182,33 @@ extract_preprocessor.workflow <- function(x, ...) {
     return(x$pre$actions$variables$variables)
   }
   abort("The workflow does not have a preprocessor.")
+}
+
+#' @export
+#' @rdname extract-workflow
+extract_parameter_set_dials.workflow <- function(x, ...) {
+  model <- extract_spec_parsnip(x)
+  param_data <- extract_parameter_set_dials(model)
+
+  if (has_preprocessor_recipe(x)) {
+    recipe <- extract_preprocessor(x)
+    recipe_param_data <- extract_parameter_set_dials(recipe)
+
+    param_data <- vctrs::vec_rbind(param_data, recipe_param_data)
+  }
+
+  dials::parameters_constr(
+    param_data$name,
+    param_data$id,
+    param_data$source,
+    param_data$component,
+    param_data$component_id,
+    param_data$object
+  )
+}
+
+#' @export
+#' @rdname extract-workflow
+extract_parameter_dials.workflow <- function(x, parameter, ...) {
+  extract_parameter_dials(extract_parameter_set_dials(x), parameter)
 }