+++ R/predict.R
@@ -29,8 +29,8 @@
 #' library(recipes)
 #' library(magrittr)
 #'
-#' training <- mtcars[1:20,]
-#' testing <- mtcars[21:32,]
+#' training <- mtcars[1:20, ]
+#' testing <- mtcars[21:32, ]
 #'
 #' model <- linear_reg() %>%
 #'   set_engine("lm")
@@ -52,7 +52,10 @@ predict.workflow <- function(object, new_data, type = NULL, opts = list(), ...) 
   workflow <- object
 
   if (!is_trained_workflow(workflow)) {
-    abort("Workflow has not yet been trained. Do you need to call `fit()`?")
+    abort(c(
+      "Can't predict on an untrained workflow.",
+      i = "Do you need to call `fit()`?"
+    ))
   }
 
   fit <- extract_fit_parsnip(workflow)