+++ tests/testthat/test-extract.R
@@ -1,3 +1,5 @@
+data(Chicago, package = "modeldata")
+
 # ------------------------------------------------------------------------------
 # extract_preprocessor()
 
@@ -36,17 +38,11 @@ test_that("can extract a variables preprocessor", {
 })
 
 test_that("error if no preprocessor", {
-  expect_error(
-    extract_preprocessor(workflow()),
-    "does not have a preprocessor"
-  )
+  expect_snapshot(error = TRUE, extract_preprocessor(workflow()))
 })
 
 test_that("error if not a workflow", {
-  expect_error(
-    extract_preprocessor(1),
-    "no applicable method"
-  )
+  expect_snapshot(error = TRUE, extract_preprocessor(1))
 })
 
 # ------------------------------------------------------------------------------
@@ -65,17 +61,11 @@ test_that("can extract a model spec", {
 })
 
 test_that("error if no spec", {
-  expect_error(
-    extract_spec_parsnip(workflow()),
-    "does not have a model spec"
-  )
+  expect_snapshot(error = TRUE, extract_spec_parsnip(workflow()))
 })
 
 test_that("error if not a workflow", {
-  expect_error(
-    extract_spec_parsnip(1),
-    "no applicable method for"
-  )
+  expect_snapshot(error = TRUE, extract_spec_parsnip(1))
 })
 
 # ------------------------------------------------------------------------------
@@ -98,17 +88,11 @@ test_that("can extract a parsnip model fit", {
 })
 
 test_that("error if no parsnip fit", {
-  expect_error(
-    extract_fit_parsnip(workflow()),
-    "does not have a model fit. Have you called `fit[(][)]` yet?"
-  )
+  expect_snapshot(error = TRUE, extract_fit_parsnip(workflow()))
 })
 
 test_that("error if not a workflow", {
-  expect_error(
-    extract_fit_parsnip(1),
-    "no applicable method for"
-  )
+  expect_snapshot(error = TRUE, extract_fit_parsnip(1))
 })
 
 # ------------------------------------------------------------------------------
@@ -152,17 +136,11 @@ test_that("can extract a mold", {
 })
 
 test_that("error if no mold", {
-  expect_error(
-    extract_mold(workflow()),
-    "does not have a mold. Have you called `fit[(][)]` yet?"
-  )
+  expect_snapshot(error = TRUE, extract_mold(workflow()))
 })
 
 test_that("error if not a workflow", {
-  expect_error(
-    extract_mold(1),
-    "no applicable method"
-  )
+  expect_snapshot(error = TRUE, extract_mold(1))
 })
 
 # ------------------------------------------------------------------------------
@@ -188,18 +166,11 @@ test_that("can extract a prepped recipe", {
   )
 
   expect_snapshot(error = TRUE, extract_recipe(workflow, FALSE))
-
-  expect_error(
-    extract_recipe(workflow, estimated = "yes please"),
-    "`estimated` must be a single `TRUE` or `FALSE`"
-  )
+  expect_snapshot(error = TRUE, extract_recipe(workflow, estimated = "yes please"))
 })
 
 test_that("error if no recipe preprocessor", {
-  expect_error(
-    extract_recipe(workflow()),
-    "must have a recipe preprocessor"
-  )
+  expect_snapshot(error = TRUE, extract_recipe(workflow()))
 })
 
 test_that("error if no mold", {
@@ -208,15 +179,171 @@ test_that("error if no mold", {
   workflow <- workflow()
   workflow <- add_recipe(workflow, recipe)
 
-  expect_error(
-    extract_recipe(workflow),
-    "does not have a mold. Have you called `fit[(][)]` yet?"
-  )
+  expect_snapshot(error = TRUE, extract_recipe(workflow))
 })
 
 test_that("error if not a workflow", {
-  expect_error(
-    extract_recipe(1),
-    "no applicable method for"
+  expect_snapshot(error = TRUE, extract_recipe(1))
+})
+
+# ------------------------------------------------------------------------------
+# extract_parameter_set_dials()
+
+test_that("extract parameter set from workflow with tunable recipe", {
+  spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(),
+                             neighbors = hardhat::tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = hardhat::tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(),
+                     deg_free = hardhat::tune(), degree = hardhat::tune())
+  lm_model <- parsnip::linear_reg() %>%
+    parsnip::set_engine("lm")
+  wf_tunable_recipe <- workflow(spline_rec, lm_model)
+
+  wf_info <- extract_parameter_set_dials(wf_tunable_recipe)
+  check_parameter_set_tibble(wf_info)
+  expect_true(all(wf_info$source == "recipe"))
+})
+
+test_that("extract parameter set from workflow with tunable model", {
+  rm_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_rm(date, ends_with("away"))
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = hardhat::tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = hardhat::tune(), noGlobalPruning = TRUE)
+  wf_tunable_model <- workflow(rm_rec, bst_model)
+
+  wf_info <- extract_parameter_set_dials(wf_tunable_model)
+  check_parameter_set_tibble(wf_info)
+  expect_equal(nrow(wf_info), 2)
+  expect_true(all(wf_info$source == "model_spec"))
+})
+
+test_that("extract parameter set from workflow with tunable recipe and model", {
+  spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(),
+                             neighbors = hardhat::tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = hardhat::tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(),
+                     deg_free = hardhat::tune(), degree = hardhat::tune())
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = hardhat::tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = hardhat::tune(), noGlobalPruning = TRUE)
+  wf_tunable <- workflow(spline_rec, bst_model)
+
+  wf_info <- extract_parameter_set_dials(wf_tunable)
+  check_parameter_set_tibble(wf_info)
+  expect_equal(
+    wf_info$source,
+    c(rep("model_spec", 2), rep("recipe", 4))
+  )
+})
+
+
+# ------------------------------------------------------------------------------
+# extract_parameter_dials()
+
+test_that("extract single parameter from workflow with tunable recipe", {
+  spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(),
+                             neighbors = hardhat::tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = hardhat::tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(),
+                     deg_free = hardhat::tune(), degree = hardhat::tune())
+  lm_model <- parsnip::linear_reg() %>%
+    parsnip::set_engine("lm")
+  wf_tunable_recipe <- workflow(spline_rec, lm_model)
+
+  expect_equal(
+    extract_parameter_dials(wf_tunable_recipe, "imputation"),
+    dials::neighbors()
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable_recipe, "threshold"),
+    dials::threshold(c(0, 1/10))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable_recipe, "deg_free"),
+    dials::spline_degree(range = c(1, 15))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable_recipe, "degree"),
+    dials::degree_int(c(1, 2))
+  )
+})
+
+test_that("extract single parameter from workflow with tunable model", {
+  rm_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_rm(date, ends_with("away"))
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = hardhat::tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = hardhat::tune(), noGlobalPruning = TRUE)
+  wf_tunable_model <- workflow(rm_rec, bst_model)
+
+  expect_equal(
+    hardhat::extract_parameter_dials(wf_tunable_model, parameter = "funky name \n"),
+    dials::trees(c(1, 100))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable_model, parameter = "rules"),
+    NA
+  )
+})
+
+test_that("extract single parameter from workflow with tunable recipe and model", {
+  spline_rec <- recipes::recipe(ridership ~ ., data = head(Chicago)) %>%
+    recipes::step_date(date) %>%
+    recipes::step_holiday(date) %>%
+    recipes::step_rm(date, ends_with("away")) %>%
+    recipes::step_impute_knn(recipes::all_predictors(),
+                             neighbors = hardhat::tune("imputation")) %>%
+    recipes::step_other(recipes::all_nominal(), threshold = hardhat::tune()) %>%
+    recipes::step_dummy(recipes::all_nominal()) %>%
+    recipes::step_normalize(recipes::all_predictors()) %>%
+    recipes::step_bs(recipes::all_predictors(),
+                     deg_free = hardhat::tune(), degree = hardhat::tune())
+  bst_model <-
+    parsnip::boost_tree(mode = "classification", trees = hardhat::tune("funky name \n")) %>%
+    parsnip::set_engine("C5.0", rules = hardhat::tune(), noGlobalPruning = TRUE)
+  wf_tunable <- workflow(spline_rec, bst_model)
+
+  expect_equal(
+    extract_parameter_dials(wf_tunable, "imputation"),
+    dials::neighbors()
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable, "threshold"),
+    dials::threshold(c(0, 1/10))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable, "deg_free"),
+    dials::spline_degree(range = c(1, 15))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable, "degree"),
+    dials::degree_int(c(1, 2))
+  )
+  expect_equal(
+    hardhat::extract_parameter_dials(wf_tunable, parameter = "funky name \n"),
+    dials::trees(c(1, 100))
+  )
+  expect_equal(
+    extract_parameter_dials(wf_tunable, parameter = "rules"),
+    NA
   )
 })