+++ NEWS.md
@@ -1,3 +1,13 @@
+# workflows 0.2.5
+
+* Improved error message in `workflow_variables()` if either `outcomes` or
+  `predictors` are missing (#144).
+
+* Removed ellipsis dependency in favor of equivalent functions in rlang.
+
+* New `extract_parameter_set_dials()` and `extract_parameter_dials()` methods 
+  to extract parameter sets and single parameters from `workflow` objects.
+
 # workflows 0.2.4
 
 * `add_model()` and `update_model()` now use `...` to separate the required